Faster cache retrieval, better use of cache and workspace
Use cache name

    cache-{{ .Branch }}-{{ arch }}-{{ checksum "Pipfile.lock" }}

in theory this should create a cache only when these factors change and that's
the only time it needs to change.

Use workspaces to save the checkout of the source files. I have no idea how code was ever
available to subsequent jobs if there was no explicit
checkout/attach_workspace/restore_cache.

Arrange cache and workspace so their paths do not intermingle and there is no
uncertainty about which saved thing is in use.

NFC: yamllint reports bad indents in last section workflows. Fix that. (not ideal that
     two things are getting changed in this commit :-/ )